+++ src/Exception/ImmutabilityException.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Headsnet\Collections\Exception;
-
-final class ImmutabilityException extends AbstractCollectionException
-{
-}