+++ src/ImmutableCollection.php
@@ -4,23 +4,14 @@ declare(strict_types=1);
 
 namespace Headsnet\Collections;
 
-use ArrayAccess;
 use Countable;
 use IteratorAggregate;
 
 /**
- * @template TKey
  * @template TValue
- * @extends IteratorAggregate<TKey, TValue>
- * @extends ArrayAccess<TKey, TValue>
+ * @extends IteratorAggregate<TValue>
+ * @extends Collection<TValue>
  */
-interface ImmutableCollection extends Countable, IteratorAggregate, ArrayAccess
+interface ImmutableCollection extends Countable, IteratorAggregate, Collection
 {
-    public function getItemClassName(): string;
-
-    /**
-     * @param TKey $index
-     * @return TValue
-     */
-    public function getItem($index);
 }